requirements.txt
peewee
mysql
mysql-connector
mysqlclient
pillow
requests
beautifulsoup4